+++ shared/api/nv7single/listui.ts
@@ -1,0 +1,7 @@
+import { OptionsItem } from "../../elem";
+import { Nv7SingleAPI } from "./nv7single";
+import { PackInfo } from "./types";
+
+export function  listUI(api: Nv7SingleAPI): OptionsItem[] {
+  return [];
+}
